Fix: WiFi client mode #47
Trying to hijack a "by default proto" in a wifi mode that could be used
for other protos breack the architecture, at current time there is no
staisfactory proposal on how to _suggest_ a default proto without
breaking the architecture.
The cleaner approach seems to have the wifi client mode working the same
as all other modes.